+++ js/libs/provoda/Model/markNestingParticipation.js
@@ -1,7 +1,7 @@
 define(function(require){
 'use strict';
-var spv = require('spv');
 var handleAdding = require('./handleNestPAdding');
+var handleRemoving = require('./handleNestPRemoving');
 
 return function markNestingParticipation(md, nesting_name, added, removed) {
   if (removed) {
@@ -24,22 +24,5 @@ return function markNestingParticipation(md, nesting_name, added, removed) {
     }
   }
 };
-
-
-function handleRemoving(md, nesting_name, item, pos) {
-  unmark(md, nesting_name, item, pos);
-}
-
-function unmark(md, nesting_name, cur) {
-  if (!cur._provoda_id) {
-    return;
-  }
-
-  var key = nesting_name + ' - ' + cur._provoda_id;
-  return spv.set.remove(cur._participation_in_nesting, key);
-}
-
-
-
 
 });
